// src/sim/data/info.gen.ts
export const MT = {
  PLAYER: 0,
  POSSESSED: 1,
  SHOTGUY: 2,
  VILE: 3,
  FIRE: 4,
  UNDEAD: 5,
  TRACER: 6,
  SMOKE: 7,
  FATSO: 8,
  FATSHOT: 9,
  CHAINGUY: 10,
  TROOP: 11,
  SERGEANT: 12,
  SHADOWS: 13,
  HEAD: 14,
  BRUISER: 15,
  BRUISERSHOT: 16,
  KNIGHT: 17,
  SKULL: 18,
  SPIDER: 19,
  BABY: 20,
  CYBORG: 21,
  PAIN: 22,
  WOLFSS: 23,
  KEEN: 24,
  BOSSBRAIN: 25,
  BOSSSPIT: 26,
  BOSSTARGET: 27,
  SPAWNSHOT: 28,
  SPAWNFIRE: 29,
  BARREL: 30,
  TROOPSHOT: 31,
  HEADSHOT: 32,
  ROCKET: 33,
  PLASMA: 34,
  BFG: 35,
  ARACHPLAZ: 36,
  PUFF: 37,
  BLOOD: 38,
  TFOG: 39,
  IFOG: 40,
  TELEPORTMAN: 41,
  EXTRABFG: 42,
  MISC0: 43,
  MISC1: 44,
  MISC2: 45,
  MISC3: 46,
  MISC4: 47,
  MISC5: 48,
  MISC6: 49,
  MISC7: 50,
  MISC8: 51,
  MISC9: 52,
  MISC10: 53,
  MISC11: 54,
  MISC12: 55,
  INV: 56,
  MISC13: 57,
  INS: 58,
  MISC14: 59,
  MISC15: 60,
  MISC16: 61,
  MEGA: 62,
  CLIP: 63,
  MISC17: 64,
  MISC18: 65,
  MISC19: 66,
  MISC20: 67,
  MISC21: 68,
  MISC22: 69,
  MISC23: 70,
  MISC24: 71,
  MISC25: 72,
  CHAINGUN: 73,
  MISC26: 74,
  MISC27: 75,
  MISC28: 76,
  SHOTGUN: 77,
  SUPERSHOTGUN: 78,
  MISC29: 79,
  MISC30: 80,
  MISC31: 81,
  MISC32: 82,
  MISC33: 83,
  MISC34: 84,
  MISC35: 85,
  MISC36: 86,
  MISC37: 87,
  MISC38: 88,
  MISC39: 89,
  MISC40: 90,
  MISC41: 91,
  MISC42: 92,
  MISC43: 93,
  MISC44: 94,
  MISC45: 95,
  MISC46: 96,
  MISC47: 97,
  MISC48: 98,
  MISC49: 99,
  MISC50: 100,
  MISC51: 101,
  MISC52: 102,
  MISC53: 103,
  MISC54: 104,
  MISC55: 105,
  MISC56: 106,
  MISC57: 107,
  MISC58: 108,
  MISC59: 109,
  MISC60: 110,
  MISC61: 111,
  MISC62: 112,
  MISC63: 113,
  MISC64: 114,
  MISC65: 115,
  MISC66: 116,
  MISC67: 117,
  MISC68: 118,
  MISC69: 119,
  MISC70: 120,
  MISC71: 121,
  MISC72: 122,
  MISC73: 123,
  MISC74: 124,
  MISC75: 125,
  MISC76: 126,
  MISC77: 127,
  MISC78: 128,
  MISC79: 129,
  MISC80: 130,
  MISC81: 131,
  MISC82: 132,
  MISC83: 133,
  MISC84: 134,
  MISC85: 135,
  MISC86: 136,
  NUMMOBJTYPES: 137,
} as const;

// mobj flags (prefix MF_ stripped).
export const MF = {
  SPECIAL: 1,
  SOLID: 2,
  SHOOTABLE: 4,
  NOSECTOR: 8,
  NOBLOCKMAP: 16,
  AMBUSH: 32,
  JUSTHIT: 64,
  JUSTATTACKED: 128,
  SPAWNCEILING: 256,
  NOGRAVITY: 512,
  DROPOFF: 1024,
  PICKUP: 2048,
  NOCLIP: 4096,
  SLIDE: 8192,
  FLOAT: 16384,
  TELEPORT: 32768,
  MISSILE: 65536,
  DROPPED: 131072,
  SHADOW: 262144,
  NOBLOOD: 524288,
  CORPSE: 1048576,
  INFLOAT: 2097152,
  COUNTKILL: 4194304,
  COUNTITEM: 8388608,
  SKULLFLY: 16777216,
  NOTDMATCH: 33554432,
  TRANSLATION: 201326592,
  TRANSSHIFT: 26,
} as const;
// ...
export type StateRow = readonly [
  sprite: number, frame: number, tics: number, action: string | null,
  nextstate: number, misc1: number, misc2: number,
];